+++ android-app/app/src/main/java/com/echopen/asso/echopen/ArchivesFragment.java
@@ -1,7 +1,7 @@
 package com.echopen.asso.echopen;
 
-import android.os.Bundle;
 import android.app.Fragment;
+import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
