+++ android-app/app/src/main/res/layout/fragment_archives.xml
@@ -14,4 +14,10 @@
         android:layout_marginTop="233dp"
         android:text="Archives"
         android:textSize="36sp" />
+
+    <ListView
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_alignParentEnd="true"
+        android:layout_alignParentTop="true" />
 </RelativeLayout>
