+++ android-app/app/src/main/res/layout/fragment_scanner.xml
@@ -8,4 +8,65 @@
         android:id="@+id/imageView"
         android:layout_width="match_parent"
         android:layout_height="match_parent"/>
+    <Button
+        android:id="@+id/button2"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:background="@drawable/scan"
+        android:layout_alignParentBottom="true"
+        android:layout_centerHorizontal="true" />
+
+    <SeekBar
+        android:id="@+id/seekBarV"
+        style="@style/Widget.AppCompat.SeekBar.Discrete"
+        android:layout_width="260dp"
+        android:layout_height="wrap_content"
+        android:text="Gain"
+        android:max="4"
+        android:progress="0"
+        android:rotation="270"
+        android:progressDrawable="@drawable/scrubber_progress"
+        android:thumb="@drawable/curseur_jauge"
+        android:visibility="invisible"
+        android:layout_marginBottom="176dp"
+        android:layout_marginRight="-100dp"
+        android:layout_above="@+id/button2"
+        android:layout_alignParentEnd="true" />
+
+    <SeekBar
+        android:id="@+id/seekBarH"
+        style="@style/Widget.AppCompat.SeekBar.Discrete"
+        android:layout_width="180dp"
+        android:text="Fréquence"
+        android:layout_height="wrap_content"
+        android:visibility="invisible"
+        android:progressDrawable="@drawable/scrubber_progress"
+        android:thumb="@drawable/curseur_jauge"
+        android:max="3"
+        android:progress="0"
+        android:layout_marginTop="36dp"
+        android:layout_alignParentTop="true"
+        android:layout_centerHorizontal="true" />
+
+    <TextView
+        android:id="@+id/frequence"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Fréquence"
+        android:visibility="invisible"
+        android:layout_marginBottom="-25dp"
+        android:layout_above="@+id/seekBarH"
+        android:layout_alignEnd="@+id/button2" />
+
+    <TextView
+        android:id="@+id/gain"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginBottom="120dp"
+        android:text="Gain"
+        android:visibility="invisible"
+        android:layout_above="@+id/seekBarV"
+        android:layout_alignParentEnd="true"
+        android:layout_marginEnd="12dp"
+        android:textColor="@color/white"/>
 </RelativeLayout>
