+++ android-app/app/src/main/res/values/styles.xml
@@ -8,5 +8,6 @@
         <item name="colorPrimary">@color/colorPrimary</item>
         <item name="colorPrimaryDark">@color/colorPrimaryDark</item>
         <item name="colorAccent">@color/colorAccent</item>
+        <item name="android:fontFamily">Roboto</item>
     </style>
 </resources>
